cloud-hypervisor: fix crash on aarch64
When running on aarch64 system cloud-hypervisor tries to call the
set_shm_regions function in the gpu class and it results in crash
because this function is not implemented. With that change we are
returning a not supported error instead of panicking so that
cloud-hypervisor is able to boot the guest virtual machine. This
is a temporary solution. In the future we will need to add support
for memory regions in the gpu device.

Signed-off-by: Yuri Nesterov <yuri.nesterov@gmail.com>